Correctly handle cases with overlapping segments
In the case where there is more than one overlapping segment, throw an
exception. I may add an ability to forcibly delete overlapping
segments in the future, but for now the behavior is unintuitive enough
that I have disallowed it.

In cases where there is only a single overlapping segment, it is safe
to delete.